icengine/icengine_api/requirements.txt
aenum==2.2.4
boto3==1.16.10
botocore==1.19.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.1.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
Pillow==8.0.1
pyproj==2.6.1.post1
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sentinelhub==3.1.0
Shapely==1.7.1
six==1.15.0
tifffile==2020.9.3
urllib3==1.25.11
utm==0.6.0
Werkzeug==1.0.1
